+++ rest-reactive/src/main/java/com/example/restreactive/service/AppointmentSlotService.java
@@ -38,7 +38,7 @@ public class AppointmentSlotService {
     }
 
 
-    public Long upsertAppointmentSlot(AppointmentSlotDto appointmentSlotDto) {
+    public Integer upsertAppointmentSlot(AppointmentSlotDto appointmentSlotDto) {
         requireNonNull(appointmentSlotDto);
         requireNonNull(appointmentSlotDto.getStartTime());
         requireNonNull(appointmentSlotDto.getEndTime());
